rename prince.php to Prince.php for case-sensitive file system autoloading
